backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pydantic
google-generativeai
python-multipart
python-dotenv
